medisecure-backend/requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
sqlalchemy==2.0.15
pydantic==1.10.8
dependency-injector==4.41.0
alembic==1.11.1
psycopg2-binary==2.9.6
python-dotenv==1.0.0
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
pytest==7.3.1
pytest-cov==4.1.0